flask-template-injection/requirements.txt
flask
pycrypto
flask_sqlalchemy
pymysql